platform: rename sh68f90a to sh68f90
The A variant is identical to the base part from the firmware's side, so the
platform takes the name of the original model and covers both. Renames the
directory, its SFR header, the meson platform key and the platform doc.

Board hardware descriptions in README.md and docs/keyboards keep saying
SH68F90A, since that is the part those boards actually carry.
