+++ docs/platforms/sh68f90.md
@@ -1,4 +1,4 @@
-# SH68F90A (BYK901 / BYK916) Support 
+# SH68F90 / SH68F90A (BYK901 / BYK916) Support
 
 ## MCU
