+++ src/platform/sh68f90/diag.h
