+++ src/platform/sh68f90/interrupts.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "sh68f90a.h"
+#include "sh68f90.h"
 
 // SDCC only emits a vector slot for a handler whose prototype is visible in the
 // translation unit that defines main(), so main.c includes this instead of every
