+++ src/platform/sh68f90/power.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "sh68f90a.h"
+#include "sh68f90.h"
 
 typedef enum {
     POWERDOWN_KEEP_USB_ALIVE,
