+++ src/platform/sh68f90/usbregs.h
